+++ Cargo.toml
@@ -25,6 +25,7 @@ members = [
     "crates/perl-refactoring",
     "crates/perl-incremental-parsing",
     "crates/perl-tdd-support",
+    "crates/perl-test-must",
     "crates/perl-lsp-providers",
     "crates/perl-lsp-rename",
     "crates/perl-lsp-protocol",
@@ -384,6 +385,7 @@ perl-lsp-on-type-formatting = { path = "crates/perl-lsp-on-type-formatting", ver
 perl-lsp-formatting-types = { path = "crates/perl-lsp-formatting-types", version = "0.12.0" }
 perl-lsp-formatting = { path = "crates/perl-lsp-formatting", version = "0.12.0" }
 perl-tdd-support = { path = "crates/perl-tdd-support", version = "0.12.0" }
+perl-test-must = { path = "crates/perl-test-must" }
 perl-lsp-diagnostic-types = { path = "crates/perl-lsp-diagnostic-types", version = "0.12.0" }
 perl-lsp-diagnostics = { path = "crates/perl-lsp-diagnostics", version = "0.12.0" }
 perl-lsp-semantic-tokens = { path = "crates/perl-lsp-semantic-tokens", version = "0.12.0" }
