+++ crates/perl-test-must/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "perl-test-must"
+version = "0.12.0"
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+description = "Panic-on-failure test helpers: must, must_some, must_err"
+readme = "README.md"
+license = "MIT OR Apache-2.0"
+repository.workspace = true
+homepage.workspace = true
+publish = false
+
+[lib]
+doctest = false
+
+# Zero dependencies — that is the entire point.
+[dependencies]
+
+[lints]
+workspace = true
